requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
